:construction: Add all monoid
